v2.7.0: Global refactoring & improvements (while keeping the same functionality)
